Fooder/requirements.txt
torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.5.3
numpy>=1.19.0
requests>=2.25.1
python-dotenv>=0.19.0
yolov5>=6.0.0
Pillow>=8.0.0
